requirements.txt
anthropic>=0.40.0,<1.0.0
watchdog>=4.0.0,<7.0.0
pypdf>=4.0.0,<6.0.0
pytesseract>=0.3.10,<0.4.0
Pillow>=10.0.0,<12.0.0
PyYAML>=6.0,<7.0
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
jinja2>=3.1.0,<4.0.0
python-multipart>=0.0.20,<1.0.0
